修改课程表的ui

+++ WePeiYang/app/src/main/java/com/twt/service/rxsrc/model/read/Review.java
@@ -20,7 +20,7 @@ public class Review {
     public String title;
     public String user_name;
     public String avatar;
-    public int scores;
+    public float scores;
     public String like_count;
     public boolean liked;
     public String content;
